fridge_image/requirements.txt
aniso8601==9.0.1
click==8.0.1
Flask==2.0.1
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.2
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
